reorganize tests to be in component directory
