+++ components/activity/editor/test/d2l-activity-editor.test.js
@@ -1,4 +1,4 @@
-import '../../components/activity/editor/d2l-hc-activity-editor.js';
+import '../d2l-hc-activity-editor.js';
 import { runConstructor } from '@brightspace-ui/core/tools/constructor-test-helper.js';
 
 describe('d2l-hc-activity-editor', () => {
